-- more modifications

+++ src/main/java/demo/service/book/BookServiceMySQL.java
@@ -7,7 +7,6 @@ import demo.model.validation.Notification;
 import demo.repository.book.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -18,7 +17,7 @@ public class BookServiceMySQL  implements BookService{
     private BookRepository bookRepository;
 
     @Override
-    public Notification<Boolean> addBook(String title, String author, String genre, int stock, double price) {
+    public Notification<Boolean> addBook(String title, String author, String genre, Integer stock, Double price) {
         Book book = new BookBuilder()
                 .setTitle(title)
                 .setAuthor(author)
@@ -41,7 +40,7 @@ public class BookServiceMySQL  implements BookService{
     }
 
     @Override
-    public Notification<Boolean> updateBook(Long id, String title, String author, String genre, int stock, double price) {
+    public Notification<Boolean> updateBook(Long id, String title, String author, String genre, Integer stock, Double price) {
         Optional<Book> optionalBook = bookRepository.findById(id);
         Book book = validateOptionalBook(title, author, genre, stock, price, optionalBook);
         BookValidator bookValidator = new BookValidator(book);
@@ -58,7 +57,7 @@ public class BookServiceMySQL  implements BookService{
         return notification;
     }
 
-    private Book validateOptionalBook(String title, String author, String genre, int stock, double price, Optional<Book> optionalBook) {
+    private Book validateOptionalBook(String title, String author, String genre, Integer stock, Double price, Optional<Book> optionalBook) {
         Book book= null;
         if (optionalBook!= null && optionalBook.isPresent()){
             book = optionalBook.get();
@@ -73,8 +72,12 @@ public class BookServiceMySQL  implements BookService{
             if (genre != null && genre.length() > 0) {
                 book.setGenre(genre);
             }
-            book.setPrice(price);
-            book.setStock(stock);
+            if (price != null) {
+                book.setPrice(price);
+            }
+            if (stock != null) {
+                book.setStock(stock);
+            }
         }
         return book;
     }
@@ -82,13 +85,15 @@ public class BookServiceMySQL  implements BookService{
     @Override
     public boolean deleteBook(Long id) {
         Book book = null;
-        Optional<Book> bookOptional = bookRepository.findById(id);
-        if (bookOptional != null && bookOptional.isPresent()) {
-            book= bookOptional.get();
-        }
-        if (book != null) {
-            bookRepository.delete(book);
-            return true;
+        if (id != null) {
+            Optional<Book> bookOptional = bookRepository.findById(id);
+            if (bookOptional != null && bookOptional.isPresent()) {
+                book = bookOptional.get();
+            }
+            if (book != null) {
+                bookRepository.delete(book);
+                return true;
+            }
         }
         return false;
     }
